[restructure] move Form to separated folder
